src/main/java/com/tasks_with_medium_level_complexity/bookstore/controller/GenreController.java
package com.tasks_with_medium_level_complexity.bookstore.controller;

import java.util.List;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import com.tasks_with_medium_level_complexity.bookstore.model.Genre;
import com.tasks_with_medium_level_complexity.bookstore.repository.GenreRepository;

@RestController
@RequestMapping("/api/genres")
public class GenreController {

    @Autowired
    private GenreRepository genreRepository;

    @GetMapping
    public List<Genre> getAllGenres() {
        return genreRepository.findAll();
    }

    @GetMapping("/{id}")
    public Genre getGenreById(@PathVariable Long id) {
        return genreRepository.findById(id).orElseThrow();
    }

    @PostMapping
    public Genre createGenre(@RequestBody Genre genre) {
        return genreRepository.save(genre);
    }

    @PutMapping("/{id}")
    public Genre updateGenre(@PathVariable Long id, @RequestBody Genre updatedGenre) {
        Genre existingGenre = genreRepository.findById(id).orElseThrow();

        existingGenre.setName(updatedGenre.getName());

        return genreRepository.save(existingGenre);
    }

    @DeleteMapping("/{id}")
    public void deleteGenre(@PathVariable Long id) {
        genreRepository.deleteById(id);
    }

    @GetMapping("/name/{name}")
    public List<Genre> getGenresByName(@PathVariable String name) {
        return genreRepository.findGenresByNameContainingIgnoreCase(name);
    }
}

